♻️ refactor: 주석 제거 및 파일명 변경으로 인한 수정 및 이름변경으로 파일 삭제
