+++ apps/users/views/admin_signup_trends_view.py
@@ -17,7 +17,7 @@ from apps.users.serializers.admin_analytics_serializers import (
     AdminSignupTrendSerializer,
 )
 from apps.users.services.admin_signup_analytics_services import get_signup_trend
-from apps.users.utils.admin_analytics_utils import IntervalLiteral
+from apps.users.utils.admin_analytics_util import IntervalLiteral
 from apps.users.utils.permissions import StaffOrSuperUser
 
 AdminSignupTrendSimpleErrorSerializer = inline_serializer(
